Go head and copy the missing step definitions and paste them into the FeatureContext class
Say what you need, but not more

For the thereAreProducts() function, change the variable to count and create a for loop. Inside, create some products and put some dummy data on each one.

Why dummy data? The definition says that we need 5 products: but it doesn't say what those products are called or how much they cost, because we don't care about that for this scenario. The point is: only include details in your scenario that you actually care about.

We'll need the entity manager in a lot of places, so create a private function getEntityManager() and return {dollar_sign}this->getContainer()->get() and pass it the service name that points directly to the entity manager.

Perfect!

Back up in thereAreProducts(), add {dollar_sign}em = {dollar_sign}this->getEntityManager(); and the usual {dollar_sign}em->persist({dollar_sign}product); and an {dollar_sign}em->flush(); at the bottom. This is easy stuff now that we've got Symfony booted