Using 'I Click' to be more Natural
Go to the next method - iClick() - and update the argument to .

We want this to work just like the built-in 'I follow' function. In fact, the only reason we're not just re-using that language is that nobody talks like that: we click things.

Anyways, the built-in functionality finds the link by its text, not a CSS selector. To use the named selector, add {dollar_sign}this->getPage()->findLink(), pass it {dollar_sign}linkText and then call click(); on that. Oh heck, let's be even lazier: just say, ->clickLink(); and be done with it
This looks for a link inside of page and then clicks it.