Finally, in iShouldSeeProducts(), we're asserting that a certain number of products are shown on the page.
In other words, once we get into the Admin section, we're looking for the number of rows in the product table.

There aren't any special classes to help find this table, but there's only one on the page, so find it via the table class. Next, use assertNotNull() in case it doesn't exist. Now, use assertCount() and pass it intval({dollar_sign}count) as the first argument.

For the second argument, we need to find all of the <tr> elements inside of the table's <tbody>. Remember, once you find an element, you can search inside of it with find() or {dollar_sign}table->findAll() to return an array of elements instead of just one. And don't forget that the first argument is still css: PhpStorm is yelling at me because I like to forget this. Ok, let's try that out!

./vendor/bin/behat features/product_admin.feature:6
Debugging Failures!

Ok, it gets further but still fails. It says:

Link 'Products' not found