Logging in... in one Step! We already did all that login stuff in authentication.feature, and I'm tempted to copy and paste all of those lines to the top of this scenario: 14 lines  features/web/authentication.feature ... lines 1 - 7   And I am on '/'   When I follow 'Login'   And I fill in 'Username' with 'admin'   And I fill in 'Password' with 'admin'   And I press 'Login' ... lines 13 - 14
But, it would be pretty lame to need to put all of this at the top of pretty much every scenario. You know what would be cooler? To just say:

21 lines  features/web/product_admin.feature
... lines 1 - 6
  Given I am logged in as an admin
... lines 8 - 21

Ooo another new step definition will be needed! Rerun the test and copy the function that behat so thoughtfully provides for us. As usual, put this in FeatureContext

Using Mink, we'll do all the steps needed to login. First, go to the login page. Normally you'd say {dollar_sign}this->getSession()->visit('/login'). But don't! Instead, wrap /login in a call to {dollar_sign}this->visitPath().

This prefixes /login - which isn't a full URL - with our base URL: http://localhost:8000.

Once we're on the login page, we need to fill out the username and password fields and press the button. We could find this stuff with CSS, but the named selector is a lot easier. Say {dollar_sign}this->getPage()->findField('Username')->setValue(). Ah, let's be lazier and do this all at once with fillField(). Pass this the label for the field - Username - and the value to fill in.

But hold on: before we fill in the rest, don't we need to make sure that this user exists in the database? Absolutely, and fortunately, we already have a function that creates a user: thereIsAnAdminUserWithPassword(). Call that from our function and pass it the usual admin / admin.

Finish by filling in the password field and pressing the button. For that, there's another shortcut: instead of findButton() then press(), use pressButton('Login').

This reproduces the steps from the login scenario, so that should be it! Run it!

./vendor/bin/behat features/product_admin.feature:6
We're in great shape.