Great - copy the iAuthorProducts() function code and paste it into our handy FeatureContext class - near the other product function.
These two functions will be similar, so we should reuse the logic. Copy the internals of thereAreProducts, make a new private function createProducts(). Pass it {dollar_sign}count as an argument and also an optional User object which will be the author for those products. Now, add an if statement that says, if {dollar_sign}author is passed then, {dollar_sign}product->setAuthor().

I already have that relationship setup with in Doctrine. Great!

In thereAreProducts(), change the body of this function to {dollar_sign}this->createProducts({dollar_sign}count).
Do the same thing in iAuthorProducts() for now. Clearly, this is still not setting the author. But I want to see if it executes first and then we'll worry about setting the author.

Cool! It runs... and fails because anonymous is still shown on the page.