Who is 'I' in a Scenario? ' The question now is: how do we get the current user? The step says 'I author'. But who is 'I' in this case? In product_admin.feature.
You can see that 'I' is whomever we logged in as. We didn't specify what the username should be for that user, but whoever is logged in is who 'I' represents.

When we worked with the ls scenarios earlier, we needed to share the command output string between the steps of a scenario. In this case, we have a similar need: we need to share the user object from the step where we log in, with the step where 'I' author some products. To share data between steps, create a new private {dollar_sign}currentUser.

In iAmLoggedInAsAnAdmin(), add {dollar_sign}this->currentUser = {dollar_sign}this->thereIsAnAdminUserWithPassword(). Click to open that function. It creates the User object of course, but now we need to also make sure it returns that.

And that's it! This login step will cause the currentUser property to be set and in iAuthorProducts() we can access that and pass it into createProducts() so that each product us authored by us.

It's pretty common to want to know who is logged in, so you'll likely want to use this in your project.

And hey it even passes! Now you can continue to write scenarios in terms of actions that 'I' take and we will actually know who 'I' is.