One of the most beautiful things about Behat is having the opportunity to do behavior driven development. This is when you write the feature first, the scenarios second and then you code it. You plan the behavior, and then make it come to life.
So far... we haven't really been doing that. We have a existing site and we've been writing features and scenarios to describe how it already behaves. That's ok, and sometimes you'll do that in your real development life. But now it's time to do BDD correctly.

Step 1: Describe the Scenario

In the 'Add a new product' scenario we're describing a feature that does not exist on the site yet. We planned this scenario earlier by planning - basically imagining - what the best behavior should be.

Step 2: Execute Behat

With that fixed, let's start to bring this feature to life by running just this scenario:

./vendor/bin/behat features/product_admin.feature:19
Our mission is clear: code just enough to fix a failure, then re-execute Behat and repeat until it's all green. The first failure is from 'When I click 'New Product''. That makes sense: that link doesn't exist. But there is something else going on too. Add an 'And print last response' line and try again:

./vendor/bin/behat features/product_admin.feature:19
Of course: we also haven't logged in yet.