Easy enough! In the product/ directory create that template: new.html.twig. Extend the base layout - layout.html.twig - and add add a body block. Add a form tag and make it submit right back to this same URL with method='POST'
I am not going to use Symfony's form system for this because (a) we don't have to and (b) my only goal is to get these tests passing. If you did want to use Symfony's form system, this is when you would start doing that!

To keep this moving, I'll paste in a bunch of code that creates the three fields we're referencing in the scenario: Name, Price and Description.

The important thing is that those names match up with the label text and that each label has a for attribute that points to the id of its field. This is how Mink can find the label by text and then find its field.

At the bottom, we have a save button that matches the second to last step

Ok, try running the scenario again!

./vendor/bin/behat features/product_admin.feature:19

We got further - it filled out and submitted the form, but didn't see the 'Product Created FTW! ' flash message. Time to add form processing logic.