Add Symfony's Request object as an argument with a type hint. Inside of newAction() add a simple if ({dollar_sign}request->isMethod('POST')).
To be super lazy, what if we cheated by not saving the product and only showing that flash message? The site already has some flash messaging functionality, so add the message that the step is looking for: {dollar_sign}this->addFlash('success', 'Product created FTW! '). Finish by redirecting the user to the product page.

Run Behat:

./vendor/bin/behat features/product_admin.feature:19

It passes, even though the product isn't saved.