BDD for Fixing Bugs Don't be a big jerk and make your tests purposefully pass like this without coding the real functionality. But sometimes, you'll write a scenario and discover later that there's a bug because you forgot to test one part of the behavior. In this case I would improve my scenario before fixing the bug.
With this, the scenario won't pass unless the product actually shows up on the product list. This is BDD: add a step to the scenario, watch it fail, and then code until it passes.

To fix this failure, add {dollar_sign}product = new Product(); with code to set the name of the product. Copy that and repeat for price and description.

This is missing validation, so you would do more work than this in real life, maybe with a scenario that guarantees that validation works.

Finish this with {dollar_sign}em = {dollar_sign}this->getDoctrine()->getManager();, then persist and flush

Bug fixed! Try it:

./vendor/bin/behat features/web/product_admin.feature:19
It's green! In fact, we can go to the browser, refresh and see the Veloci-chew toy for {dollar_sign}20.