In ProductAdminController, set the author when a product is created: {dollar_sign}product->setAuthor({dollar_sign}this->getUser())
Run Behat again:

./vendor/bin/behat features/web/product_admin.feature:19
Fixed!

And that folks, is behavior driven development. It's useful, and a bucket of fun. It forces you to design the behavior of your code. But it also helps you know when you're finished. If the scenario is green, stop coding and over-perfecting things. And yes, someone will need to add a designer's touch for a really nice UI, But from a behavioral perspective, this feature does what it needs to do. So move onto what's next.