Let's talk JavaScript and the complications it can cause. Right now everything runs using the Goutte driver and background cURL requests. This means that if your behavior relies on JavaScript, it'll fail. We need to use Selenium - or Zombie.js - another supported JavaScript driver - for these scenarios.
I want this 'New Product' button to not load a whole new page, but instead open up a modal. I cheated and already did most of the work for this. Inside of list.html.twig, add the class js-add-new-product to the 'New Product' link.

This triggers some JavaScript that I have at the bottom of the template:

 67 lines  app/Resources/views/product/list.html.twig
... lines 1 - 45
{% block javascripts %}
  {{ parent() }}
  <script type='text/javascript'>
    {dollar_sign}(document).ready(function() {
      {dollar_sign}('.js-add-new-product').on('click', function(e) {
        e.preventDefault();
        var {dollar_sign}modalContentHolder = {dollar_sign}('#modal-content-holder');
          jQuery.ajax({
            'url': {dollar_sign}(this).attr('href'),
            'success': function(content) {
            {dollar_sign}modalContentHolder.find('.modal-body').html(content);
            {dollar_sign}modalContentHolder.modal();
          }
        });
      });
    });
  </script>
{% endblock %}
Next, make the new.html.twig template only return a partial page by removing the extends and Twig block tags.
This will now only be loaded via AJAX. There are cooler ways to make this all work, but for the purposes of Behat, they all face the same complications that you'll see.

Now click the 'New Product' button. It opens up in a modal and it even saves my {dollar_sign}34 Foo product. Simple!
Since we just modified our code, we should rerun our tests to make sure that everything still works. Run the new product scenario:

./vendor/bin/behat features/web/product_admin.feature:19
It works! Wait... it shouldn't! It relies on JavaScript! In reality, the test clicked to this URL, went to an ugly, but functional page, filled out the form and hit 'Save'. That's kind of ok, it did test the form's functionality. But in reality, clicking the link opens a modal, so that should happen in the test too.