Add @javascript to the top of the scenario. That changed bumped our scenario down one line. Put the new line into the terminal and run it:
./vendor/bin/behat features/web/product_admin.feature:20
Just make sure that the Selenium server is still running in the background. Watch closely. Well, don't watch that silly Firefox error. We log in, go to the products page, clicks the 'New Product' button, fill in the form fields, and hit 'Save'. It's perfect.

Waiting for things to Happen

Reality check: if I re-ran this 5 times, I bet it would fail at least once.