In ProductAdminController pretend this isn't such a fast ajax request. Add a sleep(1) to fake the time it would take to do some logic.
Re-run things:

./vendor/bin/behat features/web/product_admin.feature:20
We log in, click the button... and the browser closes! We didn't see it fill out any fields. In the terminal, it failed at:

 34 lines  features/web/product_admin.feature
... lines 1 - 20
  Scenario: Add a new product
... lines 22 - 25
    And I fill in 'Name' with 'Veloci-chew toy'
... lines 27 - 34

Because the 'Name' field wasn't found. What is this madness?

Here's the secret: if you click a link or submit a form and it causes a full page refresh, Mink and Selenium will wait for that page refresh. But, if you do something in JavaScript, Selenium does not wait. It clicks the 'New Products' link and immediately looks for the 'Name' field. If that field isn't there almost instantly, it fails. We have to make Selenium wait after clicking the link.