What about debugging in JavaScript? Change your scenario so it fails: change the Name field to look for a non-existent 'Product Name'
Run it:

./vendor/bin/behat features/web/product_admin.feature:20
Here's the problem: you can try to watch the browser, but it happens so quickly that it's hard to see what went wrong. In the terminal, the error tells us that there isn't a field called 'Product Name', but with nothing else to help. Was there an error on the page? Are we on the wrong page? Is the field calling something else? Why won't someone tell us what's going on!?

Let me show you the master debugging tool. Google for behatch contexts(https://github.com/Behatch/contexts). This is an open source library that has a bunch of useful contexts - classes like FeatureContext and MinkContext with free definitions. For example, this has a BrowserContext(https://github.com/Behatch/contexts/blob/master/src/Context/BrowserContext.php) you could bring into your project to gain a bunch of useful definitions.