Or you could even create your own DebugContext if you wanted to organize things a bit. Shorten this to 'I break'. To use this, add this language directly above the 'Product Name' step that's failing.
The 'I' part of this language is optional. Head back to the terminal to try this:

./vendor/bin/behat features/web/product_admin.feature:20
This time, the modal pops open, and the browser freezes. The terminal just says: 'Press [RETURN] to continue...' That's right: it's waiting for us to look at the page and debug the issue. Once we know what the problem is, hit enter to let it finish. This is my favorite way to debug!