Using Gherkin TableNodes
Next, I want to add a test for this 'Is published; flag: if a product is not published, it has a little x icon. If it is published it has a ✓. I want to make sure these are showing up correctly. Right now, all of the products are unpublished.

Woo! Time for a new scenario!

But this time, we can't just say 'Given 5 products exist' because we need to control the published flag. But we can use a new trick. End the line with a colon and below, build a table just like we did earlier with scenario outlines(https://knpuniversity.com/screencast/behat/scenario-outline). Give it two headers: 'name' and 'is published': I'm making these headers up: you'll see how I use them in a second. Call the first product 'Foo1' and make it published. Call the second 'Foo2' and make it not published.

Ok, keep going on the scenario

I'll stop here for now and add the missing Then line that looks for the published flag later. Try this by running only this scenario:

./vendor/bin/behat features/web/product_admin.feature:20

Copy the new function into FeatureContext.