The TableNode Object
Ah, but this looks different: it saw that the step had a table below it and passed us a TableNode object that represents the data in the table after it. Let's iterate over the object and dump each row out to see what happens.

Science! Rerun the test:

./vendor/bin/behat features/web/product_admin.feature:20
Each row is printed as an associative array using the header and the value for that row. How cool is that?