To save some time, I'll copy some code that creates Product objects. This adds some duplication to my FeatureContext - shame on me. In a real project, I want you to be a little more careful.
In this scenario, we won't give each product an author because we don't need that for what we're testing. Call setName() passing it {dollar_sign}row['name']. Then if ({dollar_sign}row['is published'] == 'yes'), add {dollar_sign}product->setIsPublished(true);.

The is published with a space in the middle is on purpose: I want a human to be able to read the scenarios. And other than super geeks like you and I, is published reads a lot better than is_published. And, 'yes' for published is more expressive than putting a 1 or 0. In FeatureContext, we translate all of that to code.

Fetch the entity manager and flush the changes at the bottom

Great!

Cool this passes! But don't get too excited: we don't have a Then statement yet.