To finish the scenario, we need a Then that's able to look for a check mark on a specific row. The check mark icon itself is an element with a fa fa-check class.
There is no built in definition to find elements inside of a specific row. So let's describe this using natural language.

Execute Behat to get that definition printed out for us:

./vendor/bin/behat features/web/product_admin.feature:20
When you're feeling really lazy, you can add a --append-snippets flag and Behat will put the definitions inside of the FeatureContext class for you:

./vendor/bin/behat features/web/product_admin.feature:20 --append-snippets