Time for a real challenge! Deleting products, it's actually a bit harder than you might think. Here comes some curve balls -- eh eh like that baseball pun?
We need a delete button to remove individual products. BDD Time! Start with the scenario.

To delete a product, we need to start with one in the database. In fact, if we start with two products, we can delete the second and check that the first is unaffected.

Add a Given that's similar to the one from the 'show published/unpublished' scenario, but with a slight difference

Since we only care about the name, we don't need to bother with adding an 'is published' row: keep things minimal! Create a product Bar and another Foo1. Man, those dinos can't wait to get a hold of such interesting product names!

This is where it gets tricky: we'll have two rows in our table that both have 'delete' buttons, but I only want to click Delete on the second row. Add another step to do that.

Then, it would be super to see a flash message that confirms that the product was deleted. Make sure Foo1 no longer appears in this list of products. And double check that Bar was not also deleted.

This is the first time we've seen But. But, it has the same functionality as And: it extends a Then, When or Given and sounds natural.

Try it! Run just this scenario:

./vendor/bin/behat features/web/product_admin.feature:42
Copy the new function into FeatureContext