Change arg1 to linkText and arg2 to rowText
This isn't the first time we've looked for a row by finding text inside of it. Let's re-use some code.

Make a new private function findRowByText() and give it a {dollar_sign}linkText argument. Now use {dollar_sign}this->findRowByText(); in the original method and also in the new definition.
Consider the row found!