Finding Links and Buttons in a Row To find the link, we don't want to use css:  is the name of the text: what a user would see on the site. Instead, use {dollar_sign}row->findLink() and pass it {dollar_sign}linkText.
I'll repeat this one more time for fun. you can find three things by their text: links, buttons and fields. Use findLink(), findButton() and findField() on the page or individual elements to drill down to find things. Add assertNotNull({dollar_sign}link, 'Could not find link '.{dollar_sign}linkText); in case something goes wrong. Finally click that link!

We haven't done any coding yet, but the scenario is done. Run it!

./vendor/bin/behat features/product_admin.feature:42
It fails... but not in the way that I expected. It says

Undefined index: is published in FeatureContext line 68.
That's happening because - this time - we don't have the 'is published' column in our table. But on line 110, we're assuming it's always there:

 254 lines  features/bootstrap/FeatureContext.php
... lines 1 - 158
  public function theFollowingProductsExist(TableNode {dollar_sign}table){
    if ({dollar_sign}row['is published'] == 'yes') {
      {dollar_sign}product->setIsPublished(true);

... lines 171 - 175
  }

That's fine: I like to start lazy and assume everything is there. When I need the steps to be more flexible, I'll add more code.