Coding the Delete
Add a new deleteAction() and a route of /admin/products/delete/{id}. Name it product_delete. We could get fancy and add an @Method annotation that say that this will only match POST or DELETE requests. Let's keep it simple for now.

And instead of adding {dollar_sign}id as an argument to deleteAction(), I'll be even lazier and type hint the Product so that Symfony queries for it for me.

Now, remove the {dollar_sign}product, flush it, set a success flash message that matches what's in the scenario and finally redirect back to the product list route.

To add the delete link, find list.html.twig and add a column called Actions. Since you should POST to delete things, add a small form in each row, instead of a link tag. Make the form point to the product_delete route and add method='POST'. And instead of having fields, it only needs a submit button whose text is 'Delete'. Add some CSS classes to make it look nice - don't get too lazy on me.

Perfect!

Try it!

./vendor/bin/behat features/product_admin.feature:42
Click/Follow Links, Press Buttons

Hmmm, it fails in the same spot:

And I click 'Delete' in the 'Foo1' row.

Either something is wrong with the way we wrote the code, there's an error on the page or we're not even on the right page. Right now, we can't tell.