Loading the Fixtures
One way to execute the fixture is by running the doctrine:fixtures:load command. I use a different method that gives me more control. Add {dollar_sign}loader = new ContainerAwareLoader() and pass it the container.

Now, point to the exact fixtures objects that you want to load. There are two methods available: loadFromDirectory() or loadFromFile(). Move up a few directories and load from src/AppBundle/DataFixtures.

That should do it!

Next, create an {dollar_sign}executor = new ORMExecutor() and pass it the entity manager.

A purger is the second argument, which you only need if you want to clear out data. We're already doing that, so I'm not going to worry about it here. Finally type {dollar_sign}executor->execute({dollar_sign}loader->getFixtures()) and pass true as the second argument.

This says to not delete the data, but to append it instead.

Ok, run search.feature:

./vendor/bin/behat features/web/search.feature
It fails for a completely different reason. Things are never boring here! This is a unique constraint violation because it's not clearing out the data before loading the fixtures. This is a funny edge case. Because the new @BeforeScenario is near the top and the other for clearing the data is lower, they're being run in that order.